fixed total time calculate, added back button to details, made city drop down box, small tweaks. more to add

+++ ServiceLearningReporting/src/main/java/org/example/servicelearningreporting/models/ActivityForm.java
@@ -26,7 +26,7 @@ public class ActivityForm {
     private LocalDate date;
     private LocalTime startTime;
     private LocalTime endTime;
-    private int totalHours;
+    private double totalHours;
     private String organizationName;
     private String contactName;
     private String city;
